Silence compiler warning in idMaterial::IsDiscrete
for some reason `sort` is a float, and newer GCC doesn't like comparing
float with enums, so just cast it to int

+++ neo/renderer/Material.h
@@ -442,7 +442,7 @@ public:
 						// necessary to prevent mutliple gui surfaces, mirrors, autosprites, and some other
 						// special effects from being combined into a single surface
 						// guis, merging sprites or other effects, mirrors and remote views are always discrete
-	bool				IsDiscrete( void ) const { return ( entityGui || gui || deform != DFRM_NONE || sort == SS_SUBVIEW ||
+	bool				IsDiscrete( void ) const { return ( entityGui || gui || deform != DFRM_NONE || (int)sort == SS_SUBVIEW ||
 												( surfaceFlags & SURF_DISCRETE ) != 0 ); }
 
 						// Normally, dmap chops each surface by every BSP boundary, then reoptimizes.
